[debug] Do not remove install-only packages (RhBug:1844533)
Running debug-restore command may result in an attempt to remove the
running kernel (in case the running kernel package is not present in the
dump file).
Newly the install-only packages are not removed, only the versions
mentioned in the dump file are marked for installation (the decision on
what versions to keep on the system is done according to
installonly_limit config option).
New option `--remove-installonly` to force removal of those versions of
install-only packages that are not present in the dump file is
introduced.

https://bugzilla.redhat.com/show_bug.cgi?id=1844533

Closes: #403
Approved by: j-mracek